chapter01/app-demo/pyproject.toml
[project]
name = "app-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "asyncio>=3.4.3",
    "py-consul>=1.6.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.116.1",
    "pytest-asyncio>=1.1.0",
    "pytest>=8.4.1",
]

